backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
python-jose
passlib[bcrypt]
bcrypt<4.0.0